+++ CHANGELOG.md
@@ -14,6 +14,12 @@ NOTE: This is the CHANGELOG for the @looker/components MONOREPO. Each package ha
 - [Design Tokens](./packages/design-tokens/CHANGELOG.md)
 - [Icons](./packages/icons/CHANGELOG.md)
 
+## [UNRELEASED]
+
+### Fixed
+
+- update DataTable to support onClick of `Link/a tag` as text inside rows.
+
 ## [0.9.23]
 
 ### Added